requirements.txt
alembic
annotated-types
anyio
async-timeout
asyncpg
attrs
bcrypt==4.0.1
beautifulsoup4
behave
Brotli
certifi
cffi
chardet
charset-normalizer
click
colorama
cryptography
cssselect
DateTime
defusedxml
dnspython
ecdsa
email_validator
exceptiongroup
execnet
fastapi
fastapi-cli
fasteners
filelock
greenlet
h11
html5lib
httpcore
httpie
httptools
httpx
idna
iniconfig
itsdangerous
Jinja2
jwt
Mako
markdown-it-py
MarkupSafe
mdurl
mechanize
multidict
orjson
outcome
packaging
parameterized
parse
parse-type
passlib
pdbp
platformdirs
playwright
pluggy
psycopg2-binary
py
pyasn1
pycparser
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
pyee
pyfiglet
Pygments
pymongo
pynose
pyotp
pyreadline3
PySocks
pystyle
pytest
pytest-asyncio
pytest-html
pytest-metadata
pytest-ordering
pytest-rerunfailures
pytest-xdist
python-dotenv
python-jose
python-multipart
pytz
PyYAML
requests
requests-toolbelt
rich
rsa
sbvirtualdisplay
selenium
seleniumbase
shellingham
six
sniffio
sortedcontainers
soupsieve
SQLAlchemy
starlette
tabcompleter
tls-client
trio
trio-websocket
typer
typing_extensions
ujson
urllib3
uuid
uvicorn
watchfiles
webencodings
websockets
wsproto
zope.interface
zstandard
